[Refactor][Xiaomeng] add models folder, extract metrics.ts.
