--- frontend/src/__types__/enum.ts
@@ -1,13 +1,0 @@
-export enum MetricsType {
-	DEPLOYMENT_FREQUENCY = "Deployment Frequency",
-	LEAD_TIME_FOR_CHANGE = "Lead Time for Change",
-	MEAN_TIME_TO_RESTORE = "Mean Time To Restore",
-	CHANGE_FAILURE_RATE = "Change Failure Rate",
-}
-export enum MetricsLevel {
-	ELITE = "ELITE",
-	HIGH = "HIGH",
-	INVALID = "INVALID",
-	LOW = "LOW",
-	MEDIUM = "MEDIUM",
-}
