+++ frontend/src/clients/metricsApis.ts
@@ -1,33 +1,16 @@
 import { createRequest } from "./createRequest";
-import { MetricsLevel } from "../__types__/enum";
+import { MetricsInfo, MetricsUnit } from "../models/metrics";
 
-export const getFourKeyMetricsUsingPost = createRequest<
-	{
-		metricsQuery: MetricsQueryRequest;
-	},
-	FourKeyMetrics
->(({ metricsQuery }) => ({
-	method: "POST",
-	url: `/api/pipeline/metrics`,
-	data: metricsQuery,
-	headers: { "Content-Type": "application/json" },
-}));
+interface PipelineStageRequest {
+	pipelineId: string;
+	stage: string;
+}
 
 export interface MetricsQueryRequest {
 	pipelineStages: PipelineStageRequest[];
 	unit: MetricsUnit;
 	startTime: number;
 	endTime: number;
-}
-
-export interface PipelineStageRequest {
-	pipelineId: string;
-	stage: string;
-}
-
-export enum MetricsUnit {
-	FORTNIGHTLY = "Fortnightly",
-	MONTHLY = "Monthly",
 }
 
 export interface FourKeyMetrics {
@@ -37,19 +20,15 @@ export interface FourKeyMetrics {
 	meanTimeToRestore: MetricsInfo;
 }
 
-export interface MetricsInfo {
-	details: Metrics[];
-	summary: MetricsSummary;
-}
+export const getFourKeyMetricsUsingPost = createRequest<
+	{
+		metricsQuery: MetricsQueryRequest;
+	},
+	FourKeyMetrics
+	>(({ metricsQuery }) => ({
+	method: "POST",
+	url: `/api/pipeline/metrics`,
+	data: metricsQuery,
+	headers: { "Content-Type": "application/json" },
+}));
 
-export interface Metrics {
-	value: number | "NaN" | undefined;
-	endTimestamp: number;
-	startTimestamp: number;
-}
-export interface MetricsSummary extends Metrics {
-	level: MetricsLevel;
-}
-export interface ValidMetric extends Metrics {
-	value: number;
-}
