+++ frontend/src/pages/dashboard/components/MetricTooltip.tsx
@@ -2,9 +2,8 @@ import { MetricInfo } from "./MetricInfo";
 import { Button, Tooltip } from "antd";
 import { InfoCircleOutlined } from "@ant-design/icons";
 import React, { FC } from "react";
-import { MetricType } from "../../../__types__/base";
 import { GRAY_13, HINT_ICON_COLOR, OVERLAY_COLOR } from "../../../constants/styles";
-import { MetricsUnit } from "../../../clients/metricsApis";
+import { MetricsUnit, MetricType } from "../../../models/metrics";
 
 export const MetricTooltip: FC<{ unit: MetricsUnit; type: MetricType }> = ({ unit, type }) => (
 	<Tooltip
